share configs, update tsconfigs
