+++ eslint.config.mjs
@@ -1,0 +1,12 @@
+import pluginJs from "@eslint/js";
+import globals from "globals";
+import tseslint from "typescript-eslint";
+
+
+export default [
+  {files: ["**/*.{ts}"]},
+  {ignores: ["tests/**/*", "dist/**/*", "node_modules/**/*"]},
+  {languageOptions: { globals: globals.browser }},
+  pluginJs.configs.recommended,
+  ...tseslint.configs.recommended,
+];
