+++ iam/.prettierrc
@@ -1,0 +1,1 @@
+../.prettierrc
