+++ iam/eslint.config.mjs
@@ -1,0 +1,1 @@
+../eslint.config.mjs
