+++ iam/jest.config.ts
@@ -1,0 +1,1 @@
+../jest.config.ts
