+++ jest.config.ts
@@ -1,0 +1,10 @@
+import type { Config } from 'jest';
+
+const config: Config = {
+  verbose: true,
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+  testPathIgnorePatterns: ['dist/', 'node_modules/']
+};
+
+export default config;
