+++ package.json
@@ -3,12 +3,14 @@
   "version": "0.1.0",
   "description": "",
   "devDependencies": {
-		"tsx": "^4.16.2",
-		"typescript": "^5.5.3"
-	},
+    "jest": "^29.7.0",
+    "tsx": "^4.16.2",
+    "typescript": "^5.5.3"
+  },
   "scripts": {
-    "build": "pnpm build",
-    "clean": "pnpm clean",
+    "build": "pnpm -r run build",
+    "clean": "rm -rf node_modules pnpm-lock.yaml && pnpm store prune && pnpm store clear",
+    "clean:purge": "pnpm run clean && pnpm store clear",
     "test": "jest"
   },
   "keywords": [],
