--- packages/core/package.json
@@ -1,59 +1,0 @@
-{
-  "name": "@forklaunch/core",
-  "version": "0.1.9",
-  "description": "forklaunch-js core package. Contains useful building blocks.",
-  "files": [
-    "dist"
-  ],
-  "types": "dist/index.d.ts",
-  "scripts": {
-    "test": "jest",
-    "build": "tsc",
-    "docs": "typedoc --out docs *",
-    "lint": "eslint . -c eslint.config.mjs",
-    "lint:fix": "eslint . -c eslint.config.mjs --fix",
-    "format": "prettier --ignore-path=.prettierignore --config .prettierrc '**/*.ts' --write"
-  },
-  "author": "Rohin Bhargava",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/forklaunch/forklaunch-js.git"
-  },
-  "bugs": {
-    "url": "https://github.com/forklaunch/forklaunch-js/issues"
-  },
-  "homepage": "https://github.com/forklaunch/forklaunch-js#readme",
-  "dependencies": {
-    "@forklaunch/validator": "^0.3.1",
-    "@mikro-orm/core": "^6.2.9",
-    "jose": "^5.6.2",
-    "redis": "^4.6.14",
-    "uuid": "^10.0.0"
-  },
-  "devDependencies": {
-    "@eslint/js": "^9.6.0",
-    "@forklaunch/common": "^0.1.2",
-    "@types/jest": "^29.5.12",
-    "@types/qs": "^6.9.15",
-    "@types/redis": "^4.0.11",
-    "@types/uuid": "^10.0.0",
-    "globals": "^15.8.0",
-    "prettier": "^3.3.2",
-    "testcontainers": "^10.10.1",
-    "ts-jest": "^29.1.5",
-    "ts-node": "^10.9.2",
-    "typedoc": "^0.26.3",
-    "typescript": "^5.5.3",
-    "typescript-eslint": "^7.15.0"
-  },
-  "directories": {
-    "test": "tests"
-  },
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "default": "./dist/index.js"
-    }
-  }
-}
