+++ packages/universal-sdk/.prettierignore
