+++ packages/validator/.prettierignore
@@ -1,2 +1,1 @@
-node_modules/
-dist/
+../../.prettierignore
