+++ packages/validator/.prettierrc
@@ -1,7 +1,1 @@
-{
-    "semi": true,
-    "trailingComma": "none",
-    "singleQuote": true,
-    "printWidth": 80,
-    "tabWidth": 2
-}
+../../.prettierrc
